+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- The `plot_geo_poly` function now has a `single_trace` keyword argument to allow for plotting multiple polyareas provided as a single trace (default to `false`).
+- The `plot_geo_poly` function now also supports generic iterables of PolyAreas and has specific methods to be used with `AbtractRegion` inputs.
+- The `plot_geo_cells` function now has a `color_contours` keyword argument to control whether the cell contours are also colored when providing a `color` vector(default to `true`).
+
+
 ## [0.5.8] - 2025-03-20
 
 ### Added